refactored models into their respective files
