+++ score-ios/Views/ListViews/HighlightView.swift
@@ -90,103 +90,9 @@ struct HighlightSectionView: View {
     }
 }
 
-// MARK: - Model
-enum Highlight: Identifiable {
-    case video(YoutubeVideo)
-    case article(Article)
-    
-    var id: String {
-        switch self {
-        case .video(let video):
-            return video.id
-        case .article(let article):
-            return article.id
-        }
-    }
-
-    var publishedAt: String {
-        switch self {
-        case .video(let video):
-            return video.publishedAt
-        case .article(let article):
-            return article.publishedAt
-        }
-    }
-}
-
-extension Highlight{
-    static let dummydata: [Highlight] = [
-        .video(
-            YoutubeVideo(
-                id: "QGHb9heJAco",
-                title: "Cornell Celebrates Coach Mike Schafer '86",
-                description: "Cornell Celebrates Coach Mike Schafer '86 Narrated by Jeremy Schaap '91.",
-                thumbnail: "https://i.ytimg.com/vi/QGHb9heJAco/hqdefault.jpg",
-                b64Thumbnail: nil,
-                url: "https://youtube.com/watch?v=QGHb9heJAco",
-                publishedAt: "2025-10-09T00:00:00Z"
-            )
-        ),
-        .article(
-            Article(
-                id: "1",
-                title: "Cornell Daily Sun Reports Historic Win",
-                summary: "Cornell’s offense shines in a big win.",
-                image: "https://snworksceo.imgix.net/cds/2f1df221-010c-4a5b-94cc-ec7a100b7aa1.sized-1000x1000.jpg?w=1000&dpr=2",
-                url: "https://cornellsun.com/article",
-                source: "Cornell Daily Sun",
-                publishedAt: "2025-10-010T00:00:00Z"
-            )
-        ),
-        .video(
-            YoutubeVideo(
-                id: "QGHb9heJAco",
-                title: "Cornell Celebrates Coach Mike Schafer '86",
-                description: "Cornell Celebrates Coach Mike Schafer '86 Narrated by Jeremy Schaap '91.",
-                thumbnail: "https://i.ytimg.com/vi/QGHb9heJAco/hqdefault.jpg",
-                b64Thumbnail: nil,
-                url: "https://youtube.com/watch?v=QGHb9heJAco",
-                publishedAt: "2025-10-08T00:00:00Z"
-            )
-        ),
-        .article(
-            Article(
-                id: "1",
-                title: "Cornell Daily Sun Reports Historic Win",
-                summary: "Cornell’s offense shines in a big win.",
-                image: "https://snworksceo.imgix.net/cds/2f1df221-010c-4a5b-94cc-ec7a100b7aa1.sized-1000x1000.jpg?w=1000&dpr=2",
-                url: "https://cornellsun.com/article",
-                source: "Cornell Daily Sun",
-                publishedAt: "2025-10-09T00:00:00Z"
-            )
-        ),
-        .video(
-            YoutubeVideo(
-                id: "QGHb9heJAco",
-                title: "Cornell Celebrates Coach Mike Schafer '86",
-                description: "Cornell Celebrates Coach Mike Schafer '86 Narrated by Jeremy Schaap '91.",
-                thumbnail: "https://i.ytimg.com/vi/QGHb9heJAco/hqdefault.jpg",
-                b64Thumbnail: nil,
-                url: "https://youtube.com/watch?v=QGHb9heJAco",
-                publishedAt: "2025-10-010T00:00:00Z"
-            )
-        ),
-        .article(
-            Article(
-                id: "1",
-                title: "Cornell Daily Sun Reports Historic Win",
-                summary: "Cornell’s offense shines in a big win.",
-                image: "https://snworksceo.imgix.net/cds/2f1df221-010c-4a5b-94cc-ec7a100b7aa1.sized-1000x1000.jpg?w=1000&dpr=2",
-                url: "https://cornellsun.com/article",
-                source: "Cornell Daily Sun",
-                publishedAt: "2025-10-010T00:00:00Z"
-            )
-        )
-    ]
-}
 
 // MARK: - Preview
 
 #Preview {
-    HighlightView(highlights: Highlight.dummydata)
+    HighlightView(highlights: Highlight.dummyData)
 }
